requirements.txt.txt
pyTelegramBotAPI
yt-dlp
pillow
requests
